rename: 폴더 이동으로 인한 경로 설정
