+++ src/app/districts/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function boardDetail() {
+  return <div>detail</div>
+}
