--- src/app/write/page.tsx
@@ -1,11 +1,0 @@
-import { useRouter } from 'next/navigation'
-
-export default function Write() {
-  return (
-    <div>
-      <p>write</p>
-      <p>test</p>
-      <p>test1</p>
-    </div>
-  )
-}
